users/serializers.py
import logging
from rest_framework import serializers

from .models import User

logger = logging.getLogger(__name__)


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'email', 'first_name', 'password', 'is_staff']
        extra_kwargs = {'password': {'write_only': True},
                        'id': {'read_only': True}
                        }

    def create(self, validated_data):
        password = validated_data.pop('password')
        validated_data['username'] = validated_data['email']
        user = User(**validated_data)
        user.set_password(password)
        user.save()
        return user
